# Conflicts:
#	cmd/platon/config.go
#	cmd/platon/main.go
#	common/vm/inner_contract.go
#	core/blockchain.go
#	core/blockchain_cache.go
#	core/cbfttypes/type.go
#	core/types/gen_tx_json.go
#	core/types/transaction.go
#	core/vm/evm.go
#	core/vm/instructions.go
#	core/vm/staking_contract.go
#	go.mod
#	p2p/dial.go
#	p2p/discover/node.go
#	p2p/discover/table.go
#	p2p/discover/udp.go
#	p2p/server.go
#	p2p/server_test.go
#	tests/vm_test_util.go
#	x/plugin/base_plugin.go
#	x/plugin/reward_plugin.go
#	x/plugin/staking_plugin.go
#	x/plugin/staking_plugin_test.go
#	x/restricting/restricting_types.go
#	x/xutil/calculate.go

+++ core/snapshotdb/snapshotdb.go
@@ -20,8 +20,12 @@ import (
 	"container/heap"
 	"errors"
 	"fmt"
+	"github.com/PlatONnetwork/PlatON-Go/rlp"
+	"github.com/syndtr/goleveldb/leveldb/storage"
+	"golang.org/x/net/context"
 	"math/big"
 	"os"
+	"sort"
 	"sync"
 
 	"github.com/PlatONnetwork/PlatON-Go/metrics"
@@ -49,24 +53,26 @@ const (
 	MaxBlockCompaction        = 10
 	MaxBlockNotCompactionSync = 10
 	MaxBlockTriggerCompaction = 200
+	MaxCommitBlock            = 150
 )
 
-//DB the main snapshotdb interface
-//  example
-//  new a recognized blockData(sync from other peer)
-//  dbInstance.NewBlock(blockNumber, parentHash, hash)
-//  dbInstance.Put(hash, kv.key, kv.value)
-//  dbInstance.Commit(hash)
+// DB the main snapshotdb interface
 //
-//  new a unrecognized blockData(a block produce by self)
-//  dbInstance.NewBlock(blockNumber, parentHash, common.ZeroHash)
-//  dbInstance.Put(hash, kv.key, kv.value)
-//  dbInstance.Flush(hash common.Hash, blockNumber *big.Int)
-//  dbInstance.Commit(hash)
-//  get a  blockData with hash
-//  dbInstance.Get(hash, key)
-//  get a  blockData without hash
-//  dbInstance.Get(common.zerohash, key)
+//	example
+//	new a recognized BlockData(sync from other peer)
+//	dbInstance.NewBlock(blockNumber, parentHash, hash)
+//	dbInstance.Put(hash, kv.key, kv.value)
+//	dbInstance.Commit(hash)
+//
+//	new a unrecognized BlockData(a block produce by self)
+//	dbInstance.NewBlock(blockNumber, parentHash, common.ZeroHash)
+//	dbInstance.Put(hash, kv.key, kv.value)
+//	dbInstance.Flush(hash common.Hash, blockNumber *big.Int)
+//	dbInstance.Commit(hash)
+//	get a  BlockData with hash
+//	dbInstance.Get(hash, key)
+//	get a  BlockData without hash
+//	dbInstance.Get(common.zerohash, key)
 type DB interface {
 	Put(hash common.Hash, key, value []byte) error
 	NewBlock(blockNumber *big.Int, parentHash common.Hash, hash common.Hash) error
@@ -85,6 +91,8 @@ type DB interface {
 	// }
 	//
 	WalkBaseDB(slice *util.Range, f func(num *big.Int, iter iterator.Iterator) error) error
+	WalkDB(num uint64, f func(baseBlock uint64, iter iterator.Iterator, blocks []rlp.RawValue) error) error
+
 	Commit(hash common.Hash) error
 
 	// Clear close db , remove all db file
@@ -109,6 +117,7 @@ type BaseDB interface {
 	DelBaseDB(key []byte) error
 	// WriteBaseDB apply the given [][2][]byte to the baseDB.
 	WriteBaseDB(kvs [][2][]byte) error
+	WriteBaseDBWithBlock(current *types.Header, blocks []BlockData) error
 	//SetCurrent use for fast sync
 	SetCurrent(highestHash common.Hash, base, height big.Int) error
 	GetCurrent() *current
@@ -145,12 +154,13 @@ type snapshotDB struct {
 
 	unCommit *unCommitBlocks
 
-	committed  []*blockData
+	committed  []*BlockData
 	commitLock sync.RWMutex
 
-	walCh     chan *blockData
-	walExitCh chan struct{}
-	walSync   sync.WaitGroup
+	walCh         chan *BlockData
+	walLoopCtx    context.Context
+	walLoopCancel context.CancelFunc
+	walSync       sync.WaitGroup
 
 	corn *cron.Cron
 
@@ -184,7 +194,7 @@ func SetDBOptions(cache int, handles int) {
 	baseDBhandles = handles
 }
 
-//Instance return the Instance of the db
+// Instance return the Instance of the db
 func Instance() DB {
 	instance.Lock()
 	defer instance.Unlock()
@@ -239,15 +249,14 @@ func open(path string, cache int, handles int, baseOnly bool) (*snapshotDB, erro
 	}
 
 	unCommitBlock := new(unCommitBlocks)
-	unCommitBlock.blocks = make(map[common.Hash]*blockData)
+	unCommitBlock.blocks = make(map[common.Hash]*BlockData)
 	db := &snapshotDB{
 		path:          path,
 		unCommit:      unCommitBlock,
-		committed:     make([]*blockData, 0),
+		committed:     make([]*BlockData, 0),
 		baseDB:        baseDB,
 		snapshotLockC: snapshotUnLock,
-		walCh:         make(chan *blockData, 2),
-		walExitCh:     make(chan struct{}),
+		walCh:         make(chan *BlockData, 2),
 	}
 	if baseOnly {
 		return db, nil
@@ -288,6 +297,67 @@ func Open(path string, cache int, handles int, baseOnly bool) (DB, error) {
 	return db, nil
 }
 
+func OpenWithStorage(st storage.Storage, cache int, handles int, baseOnly bool) (*snapshotDB, error) {
+	logger.Info("open snapshot db Allocated cache and file handles", "cache", cache, "handles", handles, "baseDB", baseOnly)
+
+	baseDB, err := leveldb.Open(st, &opt.Options{
+		OpenFilesCacheCapacity: handles,
+		BlockCacheCapacity:     cache / 2 * opt.MiB,
+		WriteBuffer:            cache / 4 * opt.MiB, // Two of these are used internally
+		Filter:                 filter.NewBloomFilter(10),
+	})
+	if err != nil {
+		if _, corrupted := err.(*leveldbError.ErrCorrupted); corrupted {
+			baseDB, err = leveldb.Recover(st, nil)
+			if err != nil {
+				return nil, fmt.Errorf("[SnapshotDB.recover]RecoverFile baseDB fail:%v", err)
+			}
+		} else {
+			return nil, err
+		}
+	}
+
+	unCommitBlock := new(unCommitBlocks)
+	unCommitBlock.blocks = make(map[common.Hash]*BlockData)
+	db := &snapshotDB{
+		path:          "",
+		unCommit:      unCommitBlock,
+		committed:     make([]*BlockData, 0),
+		baseDB:        baseDB,
+		snapshotLockC: snapshotUnLock,
+	}
+	if baseOnly {
+		return db, nil
+	}
+
+	_, getCurrentError := baseDB.Get([]byte(CurrentSet), nil)
+	if getCurrentError == nil {
+		logger.Info("begin recover", "path", "")
+		if err := db.loadCurrent(); err != nil {
+			return nil, err
+		}
+		logger.Info("load current", "base", db.current.base, "high", db.current.highest)
+		if err := db.recover(); err != nil {
+			logger.Error("recover db fail:", "error", err)
+			return nil, err
+		}
+	} else if getCurrentError == leveldb.ErrNotFound {
+		logger.Info("begin init db current", "path", "")
+		if err := db.SetCurrent(common.ZeroHash, *common.Big0, *common.Big0); err != nil {
+			return nil, err
+		}
+	} else {
+		return nil, getCurrentError
+	}
+
+	if !baseOnly {
+		if err := db.Start(); err != nil {
+			return nil, err
+		}
+	}
+	return db, nil
+}
+
 func copyDB(from, to *snapshotDB) {
 	to.path = from.path
 	to.current = from.current
@@ -297,26 +367,40 @@ func copyDB(from, to *snapshotDB) {
 	to.corn = from.corn
 	to.closed = from.closed
 	to.snapshotLockC = from.snapshotLockC
-	to.walExitCh = from.walExitCh
+	to.walLoopCtx = from.walLoopCtx
 	to.walCh = from.walCh
 }
 
 func initDB(path string, sdb *snapshotDB) error {
-	dbInterface, err := open(path, baseDBcache, baseDBhandles, false)
-	if err != nil {
-		return err
+	if path == "" {
+		// only for test
+		dbInterface, err := OpenWithStorage(storage.NewMemStorage(), baseDBcache, baseDBhandles, false)
+		if err != nil {
+			return err
+		}
+		copyDB(dbInterface, sdb)
+		if err := sdb.Start(); err != nil {
+			return err
+		}
+		return nil
+	} else {
+		dbInterface, err := open(path, baseDBcache, baseDBhandles, false)
+		if err != nil {
+			return err
+		}
+		copyDB(dbInterface, sdb)
+		if err := sdb.Start(); err != nil {
+			return err
+		}
+		return nil
 	}
-	copyDB(dbInterface, sdb)
-	if err := sdb.Start(); err != nil {
-		return err
-	}
-	return nil
 }
 
 func (s *snapshotDB) Start() error {
 	if err := s.cornStart(); err != nil {
 		return err
 	}
+	s.walLoopCtx, s.walLoopCancel = context.WithCancel(context.Background())
 	go s.loopWriteWal()
 	return nil
 }
@@ -344,6 +428,31 @@ func (s *snapshotDB) WriteBaseDB(kvs [][2][]byte) error {
 	}
 	if err := s.baseDB.Write(batch, nil); err != nil {
 		return err
+	}
+	return nil
+}
+
+func (s *snapshotDB) WriteBaseDBWithBlock(current *types.Header, blocks []BlockData) error {
+	sort.Slice(blocks, func(i, j int) bool {
+		return blocks[i].Number.Uint64() < blocks[j].Number.Uint64()
+	})
+	batch := new(leveldb.Batch)
+	for _, block := range blocks {
+		itr := block.data.NewIterator(nil)
+		for itr.Next() {
+			if itr.Value() == nil || len(itr.Value()) == 0 {
+				batch.Delete(itr.Key())
+			} else {
+				batch.Put(itr.Key(), itr.Value())
+			}
+		}
+		itr.Release()
+	}
+	if err := s.baseDB.Write(batch, nil); err != nil {
+		return fmt.Errorf("write to baseDB fail,%v", err)
+	}
+	if err := s.SetCurrent(current.Hash(), *current.Number, *current.Number); err != nil {
+		return fmt.Errorf("SetCurrent fail, %v", err)
 	}
 	return nil
 }
@@ -510,7 +619,7 @@ func (s *snapshotDB) findToWrite() int {
 
 	var length = commitNum
 	for i := 0; i < length; i++ {
-		if s.committed[i].Number.Cmp(minimumHeight) > 0 {
+		if s.committed[i].Number.Cmp(minimumHeight) >= 0 {
 			commitNum--
 		}
 	}
@@ -539,10 +648,10 @@ func (s *snapshotDB) writeToBasedb(commitNum int) error {
 	return nil
 }
 
-//NewBlock call when you need a new unRecognized or recognized  block data
-//it will set JournalHeader for the block
-//if hash nil ,new unRecognized data
-//if hash not nul,new Recognized data
+// NewBlock call when you need a new unRecognized or recognized  block data
+// it will set JournalHeader for the block
+// if hash nil ,new unRecognized data
+// if hash not nul,new Recognized data
 func (s *snapshotDB) NewBlock(blockNumber *big.Int, parentHash common.Hash, hash common.Hash) error {
 	if blockNumber == nil {
 		return errors.New("[SnapshotDB]the blockNumber must not be nil ")
@@ -551,7 +660,7 @@ func (s *snapshotDB) NewBlock(blockNumber *big.Int, parentHash common.Hash, hash
 		logger.Error("the block is less than commit highest", "commit", s.current.GetHighest(false).Num, "new", blockNumber)
 		return ErrBlockTooLow
 	}
-	block := new(blockData)
+	block := new(BlockData)
 	block.Number = new(big.Int).Set(blockNumber)
 	block.ParentHash = parentHash
 	block.BlockHash = hash
@@ -670,8 +779,8 @@ func (s *snapshotDB) GetBaseDB(key []byte) ([]byte, error) {
 	}
 }
 
-//Has check the key is exist in chain
-//same logic with get
+// Has check the key is exist in chain
+// same logic with get
 func (s *snapshotDB) Has(hash common.Hash, key []byte) (bool, error) {
 	_, err := s.Get(hash, key)
 	if err == nil {
@@ -717,7 +826,7 @@ func (s *snapshotDB) Flush(hash common.Hash, blockNumber *big.Int) error {
 	return nil
 }
 
-func (s *snapshotDB) theBlockIsCommit(block *blockData) bool {
+func (s *snapshotDB) theBlockIsCommit(block *BlockData) bool {
 	if block.Number.Cmp(s.current.GetHighest(false).Num) != 0 {
 		return false
 	}
@@ -790,7 +899,7 @@ func (s *snapshotDB) BaseNum() (*big.Int, error) {
 // content of snapshot are guaranteed to be consistent.
 // slice
 func (s *snapshotDB) WalkBaseDB(slice *util.Range, f func(num *big.Int, iter iterator.Iterator) error) error {
-	logger.Debug("begin walkbase db")
+	logger.Debug("begin walk base db")
 	snapshot, err := s.baseDB.GetSnapshot()
 	if err != nil {
 		return errors.New("[snapshotdb] get snapshot fail:" + err.Error())
@@ -812,11 +921,53 @@ func (s *snapshotDB) Clear() error {
 	if err := s.Close(); err != nil {
 		return err
 	}
-	logger.Info("begin clear file", "path", s.path)
-	if err := os.RemoveAll(s.path); err != nil {
-		return err
+	if s.path != "" {
+		logger.Info("begin clear file", "path", s.path)
+		if err := os.RemoveAll(s.path); err != nil {
+			return err
+		}
 	}
 	return nil
+}
+
+// 根据块高来迭代数据库
+func (s *snapshotDB) WalkDB(num uint64, f func(baseBlock uint64, iter iterator.Iterator, blocks []rlp.RawValue) error) error {
+	logger.Debug("begin walk db", "pivot", num)
+
+	s.commitLock.Lock()
+
+	baseBlock := s.current.GetBase(false).Num.Uint64()
+	highestBlock := s.current.GetHighest(false).Num.Uint64()
+	if num > highestBlock || num < baseBlock {
+		s.commitLock.Unlock()
+		return fmt.Errorf("the pivot block %v request seems not in commit block,highest %d,base %d", num, highestBlock, baseBlock)
+	}
+
+	blocks := make([]rlp.RawValue, 0)
+	for i := 0; i < len(s.committed); i++ {
+		if s.committed[i].Number.Uint64() > num {
+			break
+		}
+		tmp, err := rlp.EncodeToBytes(s.committed[i])
+		if err != nil {
+			s.commitLock.Unlock()
+			return fmt.Errorf("WalkDB fail , encode to bytes error,%v", err)
+		}
+		blocks = append(blocks, tmp)
+	}
+	snapshot, err := s.baseDB.GetSnapshot()
+	if err != nil {
+		s.commitLock.Unlock()
+		return fmt.Errorf("get snapshot fail,%v", err)
+	}
+	s.commitLock.Unlock()
+	defer snapshot.Release()
+	t := snapshot.NewIterator(nil, nil)
+	defer func() {
+		logger.Debug("walk DB release")
+		t.Release()
+	}()
+	return f(baseBlock, t, blocks)
 }
 
 // Ranking return iterates  of the DB.
@@ -882,7 +1033,10 @@ func (s *snapshotDB) Close() error {
 		s.corn.Stop()
 	}
 	s.walSync.Wait()
-	close(s.walExitCh)
+
+	if s.walLoopCancel != nil {
+		s.walLoopCancel()
+	}
 
 	if s.baseDB != nil {
 		if err := s.baseDB.Close(); err != nil {
